Clang: Fix GCC 5.3 warnings in unit tests
signed/unsigned comparisons.

+++ tests/unit/unittest/skippedsourcerangestest.cpp
@@ -79,9 +79,9 @@ MATCHER_P4(IsSourceLocation, filePath, line, column, offset,
            )
 {
     if (!arg.filePath().endsWith(filePath)
-     || arg.line() != line
-     || arg.column() != column
-     || arg.offset() != offset) {
+     || arg.line() != uint(line)
+     || arg.column() != uint(column)
+     || arg.offset() != uint(offset)) {
         return false;
     }
